pyproject.toml
[project]
name = "mcp-cnb-knowledge"
version = "0.1.0"
description = "CNB Knowledge Base MCP Server - Query CNB repository documentation using RAG"
authors = [
    {name = "hicaosen",email = "cscg52@qq.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastmcp (>=2.12.4,<3.0.0)",
    "httpx (>=0.28.1,<1.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[project.scripts]
mcp-cnb-knowledge = "mcp_cnb_knowledge.server:run"

[tool.poetry]
packages = [{include = "mcp_cnb_knowledge", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
